blastn clean ups

+++ src/viral_ngs/taxon_filter.py
@@ -346,15 +346,30 @@ def multi_db_deplete_bam(inBam, refDbs, deplete_method, outBam, threads=1, JVMme
 
 def run_blastn(blastn_path, db, input_fasta, blast_threads=1):
     """ run blastn on the input fasta file. this is intended to be run in parallel """
-    chunk_hits = mkstempfname('.hits.txt')
+    chunk_hits = mkstempfname('.hits.txt.gz')
+
     blastnCmd = [
         blastn_path, '-db', db, '-word_size', '16', '-num_threads', str(blast_threads), '-evalue', '1e-6', '-outfmt',
-        '6', '-max_target_seqs', '2', '-query', input_fasta, '-out', chunk_hits
+        '6', '-max_target_seqs', '1', '-query', input_fasta,
     ]
     log.debug(' '.join(blastnCmd))
-    util.misc.run_and_print(blastnCmd, check=True)
+    blast_pipe = subprocess.Popen(blastnCmd, stdout=subprocess.PIPE)
 
+    with util.file.open_or_gzopen(chunk_hits, 'wt') as outf:
+        # strip tab output to just query read ID names and emit
+        last_read_id = None
+        for line in blast_pipe.stdout:
+            line = line.decode('UTF-8').rstrip('\n\r')
+            read_id = line.split('\t')[0]
+            # only emit if it is not a duplicate of the previous read ID
+            if read_id != last_read_id:
+                last_read_id = read_id
+                outf.write(read_id + '\n')
+
+    if blast_pipe.poll():
+        raise CalledProcessError()
     os.unlink(input_fasta)
+
     return chunk_hits
 
 
@@ -442,31 +457,29 @@ def blastn_chunked_fasta(fasta, db, chunkSize=1000000, threads=1):
 
 
 def deplete_blastn_bam(inBam, db, outBam, threads, chunkSize=1000000, JVMmemory=None):
+#def deplete_blastn_bam(inBam, db, outBam, threads, chunkSize=0, JVMmemory=None):
     'Use blastn to remove reads that match at least one of the databases.'
 
-    fastq1 = mkstempfname('.1.fastq')
-    fasta = mkstempfname('.1.fasta')
     blast_hits = mkstempfname('.blast_hits.txt')
-    blastOutFile = mkstempfname('.hits.txt')
 
-    # Initial BAM -> FASTQ pair
-    tools.samtools.SamtoolsTool().bam2fq(inBam, fastq1)
+    if chunkSize:
+        ## chunk up input and perform blastn in several parallel threads
 
-    # Find BLAST hits
-    read_utils.fastq_to_fasta(fastq1, fasta)
-    os.unlink(fastq1)
-    log.info("running blastn on %s against %s", inBam, db)
-    blastOutFiles = blastn_chunked_fasta(fasta, db, chunkSize, threads)
-    with open(blast_hits, 'wt') as outf:
-        for blastOutFile in blastOutFiles:
-            with open(blastOutFile, 'rt') as inf:
-                for line in inf:
-                    idVal = line.split('\t')[0].strip()
-                    if idVal.endswith('/1') or idVal.endswith('/2'):
-                        idVal = idVal[:-2]
-                    outf.write(idVal + '\n')
-            os.unlink(blastOutFile)
-    os.unlink(fasta)
+        # Initial BAM -> FASTA sequences
+        fasta = mkstempfname('.fasta')
+        tools.samtools.SamtoolsTool().bam2fa(inBam, fasta)
+
+        # Find BLAST hits
+        log.info("running blastn on %s against %s", inBam, db)
+        blastOutFiles = blastn_chunked_fasta(fasta, db, chunkSize, threads)
+        util.file.cat(blast_hits, blastOutFiles)
+        os.unlink(fasta)
+
+    else:
+        ## pipe tools together and run blastn multithreaded
+        with open(blast_hits, 'wt') as outf:
+            for read_id in tools.blast.BlastnTool().get_hits(inBam, db, threads=threads):
+                outf.write(read_id + '\n')
 
     # Deplete BAM of hits
     tools.picard.FilterSamReadsTool().execute(inBam, True, blast_hits, outBam, JVMmemory=JVMmemory)
